centralizing the FHIRSpecificationData classes, by implementing the common functions in the parent class and inheriting them from the child classes
